feat: add data structures for AI providers and steps; modularize code for better organization and maintainability
